+++ trunk/src/argonms/login/LoginCharacter.java
@@ -207,18 +207,18 @@ public class LoginCharacter extends LimitedActionCharacter {
 		//TODO: get real equipped inventory size?
 		inventories.put(InventoryType.EQUIPPED, equipped);
 		InventoryTools.equip(equipment, equipped,
-				InventoryTools.addToInventory(equipment, top, (short) 1)
+				InventoryTools.addToInventory(equipment, top, 1)
 				.addedOrRemovedSlots.get(0).shortValue(), (short) -5);
 		InventoryTools.equip(equipment, equipped,
-				InventoryTools.addToInventory(equipment, bottom, (short) 1)
+				InventoryTools.addToInventory(equipment, bottom, 1)
 				.addedOrRemovedSlots.get(0).shortValue(), (short) -6);
 		InventoryTools.equip(equipment, equipped,
-				InventoryTools.addToInventory(equipment, shoes, (short) 1)
+				InventoryTools.addToInventory(equipment, shoes, 1)
 				.addedOrRemovedSlots.get(0).shortValue(), (short) -7);
 		InventoryTools.equip(equipment, equipped,
-				InventoryTools.addToInventory(equipment, weapon, (short) 1)
+				InventoryTools.addToInventory(equipment, weapon, 1)
 				.addedOrRemovedSlots.get(0).shortValue(), (short) -11);
-		InventoryTools.addToInventory(etc, 4161001, (short) 1);
+		InventoryTools.addToInventory(etc, 4161001, 1);
 		p.addInventories(inventories);
 
 		p.bindings = new HashMap<Byte, KeyBinding>();
